New version - 0.2
List containers take max width if not specified

ListRowContainer and ListColumnContainer could have unspecified width.
They will take maximal possible width and distribute this to the
columns.

List containers will raise an exception if it is not possible to render
container content. Or give user hint to turn off numbering if it can be
rendered without the numbering.