Add possibility to change setup.py arguments
By default we want to create an archive with bdist_wheel which will also
create metadata for PYPI distribution. However, dbist_wheel have another
dependency package which may not be by default installed and it is not
required for spec file creation at all.

Add setup.py arguments as Makefile variable and adapt Packit to change
this to remove the bdist_wheel argument.